Move check into ChoiceParameterDefinition.DescriptorImpl.
