+++ core/src/test/java/hudson/model/ChoiceParameterDefinitionTest.java
@@ -1,7 +1,10 @@
 package hudson.model;
 
+import hudson.util.FormValidation;
+import jenkins.model.Jenkins;
 import org.junit.Test;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
@@ -13,5 +16,13 @@ public class ChoiceParameterDefinitionTest {
         assertTrue(ChoiceParameterDefinition.areValidChoices("abc"));
         assertTrue(ChoiceParameterDefinition.areValidChoices("abc\ndef"));
         assertTrue(ChoiceParameterDefinition.areValidChoices("abc\r\ndef"));
+    }
+
+    @Test
+    public void testCheckChoices() throws Exception {
+        ChoiceParameterDefinition.DescriptorImpl descriptorImpl = new ChoiceParameterDefinition.DescriptorImpl();
+
+        assertEquals(FormValidation.Kind.OK, descriptorImpl.doCheckChoices("abc\ndef").kind);
+        assertEquals(FormValidation.Kind.ERROR, descriptorImpl.doCheckChoices("").kind);
     }
 }
